requirements.txt
requests>=2.32.0
pydantic>=2.7.0
python-dotenv>=1.0.1
PyQt5>=5.15.11
PyYAML>=6.0.2
tenacity>=8.2.3
